feat : 2km 반경에 있는 팀 조회 하는 쿼리 작성

+++ src/main/java/univ/kgu/carely/domain/team/dto/response/ResTeamOutlineDTO.java
@@ -22,6 +22,6 @@ public class ResTeamOutlineDTO {
     private String teamName;
     private Address address;
     private Double distance;
-    private Integer memberCount;
+    private Long memberCount;
 
 }
